[TASK] Start nginx before thin
This makes the intergration tests work.

I am not entirely sure, but the wait_until_ready!
function did not block until thin was completely
started. I assume that it is because of
nginx was not yet restarted with the thin backend
so that we were hitting the static default site.

Tests run successful, if we install nginx first
and then thin.